PWBE2/projetoEscola/back/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et-xmlfile==1.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
